refactor: confirm/alert/announcement dialog moved and share code
