--- src/components/alert-dialog/AlertDialog.module.scss
@@ -1,33 +1,0 @@
-@use 'theme';
-
-.aminoContent {
-  padding: theme.$amino-space-24;
-
-  & > div {
-    text-align: center;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-}
-
-.aminoFooter {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-
-  & > button + button {
-    margin-left: theme.$amino-space-8;
-  }
-
-  button {
-    flex: 1;
-  }
-}
-
-.aminoAlertPrompt {
-  color: theme.$amino-text-color-secondary;
-  line-height: theme.$amino-line-height-base;
-  font-size: theme.$amino-font-size-base;
-}
