--- src/components/alert-dialog/AlertDialog.module.scss.d.ts
@@ -1,3 +1,0 @@
-export declare const aminoAlertPrompt: string;
-export declare const aminoContent: string;
-export declare const aminoFooter: string;
