+++ src/components/dialog/alert/useAlert.tsx
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
 
-import { AlertContext } from 'src/components/alert-dialog/AlertContext';
+import { AlertContext } from 'src/components/dialog/alert/AlertContext';
 
 export const useAlert = () => useContext(AlertContext);
