+++ src/components/dialog/announcement/AnnouncementDialog.module.scss.d.ts
