+++ src/components/dialog/announcement/__stories__/AnnouncementDialog.stories.tsx
@@ -2,11 +2,11 @@ import { useState } from 'react';
 
 import type { Meta, StoryFn } from '@storybook/react';
 
+import { Button } from 'src/components/button/Button';
 import {
   type AnnouncementDialogProps,
   AnnouncementDialog,
-} from 'src/components/announcement-dialog/AnnouncementDialog';
-import { Button } from 'src/components/button/Button';
+} from 'src/components/dialog/announcement/AnnouncementDialog';
 import { Text } from 'src/components/text/Text';
 import { ArrowRightIcon } from 'src/icons/ArrowRightIcon';
