+++ src/components/dialog/confirm/ConfirmDialog.tsx
@@ -1,0 +1,40 @@
+import { Button } from 'src/components/button/Button';
+import {
+  type DismissableDialogProps,
+  _dismissableDialogGetButtonVariant,
+  DismissableDialog,
+} from 'src/components/dialog/DismissableDialog';
+
+export type ConfirmDialogProps = DismissableDialogProps & {
+  confirmText: string;
+  dismissText: string;
+  confirmAction: () => void;
+  dismissAction: () => void;
+};
+
+export const ConfirmDialog = ({
+  confirmAction,
+  confirmText,
+  dismissAction,
+  dismissText,
+  ...props
+}: ConfirmDialogProps) => (
+  <DismissableDialog
+    {...props}
+    actions={
+      <>
+        <Button onClick={dismissAction} size="lg">
+          {dismissText}
+        </Button>
+        <Button
+          onClick={confirmAction}
+          size="lg"
+          variant={_dismissableDialogGetButtonVariant(props.intent)}
+        >
+          {confirmText}
+        </Button>
+      </>
+    }
+    dismissAction={dismissAction}
+  />
+);
