+++ src/components/dialog/confirm/useConfirm.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
 
-import { ConfirmContext } from 'src/components/confirm-dialog/ConfirmContext';
+import { ConfirmContext } from 'src/components/dialog/confirm/ConfirmContext';
 
 export const useConfirm = () => useContext(ConfirmContext);
